Update License.yaml (#176)
The schema is currently forcing the developer to select a public license from [the list documented by GitHub](https://docs.github.com/en/github/creating-cloning-and-archiving-repositories/licensing-a-repository), but the developer may not want to specify 1) one of these licenses or 2) a license at all. The solution to 1) is for the developer to contact us so we add the license to the list. This PR aims to solve 2).